pyproject.toml
[project]
name = "legaladvisorsystem"
version = "0.1.0"
description = "Legal Advisor System with 2026 Security Patches"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "colorlog>=6.10.1",
    "docling>=2.74.0",
    "docling-core>=2.65.1",
    "fastembed>=0.7.4",
    "flashrank>=0.2.10",
    "hf-xet>=1.2.0",
    "huggingface-hub>=0.36.2",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-docling>=2.0.0",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-postgres>=3.0.4",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.3.3",
    "psycopg-pool>=3.3.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "tqdm>=4.67.3",
    "transformers>=4.57.6",
    "pillow>=11.2.1,<12.0",
    "torch>=2.10.0",  
    "torchvision>=0.25.0", 
    "torchaudio>=2.10.0"
]

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch-cu126" # Updated to cu126
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
torchaudio = { index = "pytorch-cu126" }